js/xpconnect/tests/mochitest/test_bug1448048.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1448048
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1448048</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

    /** Test for Bug 1448048 **/
    var { AppConstants } = SpecialPowers.ChromeUtils.importESModule(
      "resource://gre/modules/AppConstants.sys.mjs"
    );
    if (AppConstants.NIGHTLY_BUILD) {
      is(typeof Components, "undefined", "Should be no Components shim on Nightly");
    } else {
      is(typeof Components, "object", "Should have a components shim on non-Nightly");
    }
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1448048">Mozilla Bug 1448048</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
